chore: improved help
