+++ src/cli/commands/mc-version-help.ts
@@ -1,8 +1,7 @@
 import { CommanderStatic } from "commander";
 import { log } from "console";
-import * as updateNotifier from "update-notifier";
-import { MC_NAME, MC_VERSION } from "../../version";
-import { getColor } from "./command-utils";
+import { MC_VERSION } from "../../version";
+import { checkForUpdates, getColor } from "./command-utils";
 
 const magenta = getColor("magenta");
 const cyan = getColor("cyan");
@@ -28,23 +27,7 @@ export default (program: CommanderStatic) => {
         log(`    the credentials and cookies should only be used in secure environments`);
         log(`    Full documentation: ${cyan("https://opensource.mindsphere.io")}\n`);
 
-        const pkgInfo = {
-            pkg: {
-                name: `@mindconnect/${MC_NAME}`,
-                version: `${MC_VERSION}`,
-            },
-        };
-
-        const notifier = updateNotifier(pkgInfo);
-
-        if (notifier.update) {
-            console.log(
-                `\n\t There is an update available: ${magenta(notifier.update.latest + " ")} (${notifier.update.type})`
-            );
-            console.log(`\t Run ${magenta("npm install -g ")}${magenta(pkgInfo.pkg.name)} to update`);
-            console.log(`\t or download the release for your system from`);
-            console.log(`\t ${magenta("https://github.com/mindsphere/mindconnect-nodejs/releases")}\n`);
-        }
+        checkForUpdates();
     });
 
     program.on("command:*", function () {
